requirements.txt
django-cms>=3.4.5,<4.1